Remove long/double type from core module test cases.

+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/values/BFloat.java
@@ -62,6 +62,6 @@ public final class BFloat extends BValueType {
 
     @Override
     public boolean equals(Object obj) {
-        return ((BFloat) obj).floatValue() == value;
+        return ((BFloat) obj).doubleValue() == value;
     }
 }
